+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Hydrograph can now show either last 7 days, last 30 days, or the last year of data for
+the selected timeseries.
+
+
 ### Changed
 - Cursor location / tooltip defaults to last point in the time series.
 - Date labels moved to and centered in areas between midnight tick marks